pyproject.toml
[project]
name = "deberta-cls"
version = "0.1.0"
description = "repo for mlops project at mipt with deberta classification"
authors = [
    {name = "mdr333@mail.ru"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy (>=2.2.3,<3.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "torch (>=2.6.0,<3.0.0)",
    "lightning (>=2.5.0.post0,<3.0.0)",
    "datasets (>=3.3.2,<4.0.0)",
    "transformers (>=4.49.0,<5.0.0)",
    "pre-commit (>=4.1.0,<5.0.0)",
    "hydra-core (>=1.3.2,<2.0.0)",
    "accelerate (>=1.5.2,<2.0.0)",
    "mlflow (>=2.21.3,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
